reusable functions moved to common.js
